Introduce following values mixin
